# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MyLocalTrade API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /conversations/{id}/bookings:
    post:
      operationId: proposeBooking
      tags: [bookings]
      summary: Propose an appointment (either party, hired conversations only). A new proposal supersedes any existing live booking.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking proposed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "400":
          description: Invalid date/time (must be in the future)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Only participants of this conversation can propose a booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conversation is not in a hired state, or is cancelled/completed/closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /conversations/{id}/booking-slots:
    get:
      tags: [bookings]
      summary: Available appointment start times for the conversation's trader on a UK-local date.
      operationId: getBookingSlots
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          required: true
          schema:
            type: string
          description: UK-local date, YYYY-MM-DD.
        - name: durationMinutes
          in: query
          required: false
          schema:
            type: integer
            enum: [30, 60, 90, 120, 180, 240, 480]
      responses:
        "200":
          description: Available start times
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingSlotsResponse"
        "400":
          description: Invalid date or duration
        "404":
          description: Conversation not found or caller is not a participant
  /bookings/{id}/confirm:
    post:
      operationId: confirmBooking
      tags: [bookings]
      summary: Confirm a proposed appointment (only the party who did NOT propose it)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Only the other party can confirm a proposed booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Booking is no longer proposed (confirmed, cancelled or superseded)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}/cancel:
    post:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a proposed or confirmed appointment (either party)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Only participants of this conversation can cancel the booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Booking is not live (already cancelled or superseded)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Booking:
      type: object
      description: |
        A lightweight appointment inside a hired conversation. Times are UTC
        instants; clients render them in the user's local timezone (UK).
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        startAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
          description: Appointment length in minutes (legacy bookings report 60).
        endAt:
          type: string
          format: date-time
          description: Reserved interval end (startAt + duration).
        note:
          type: string
          nullable: true
        status:
          type: string
          enum: [PROPOSED, CONFIRMED, CANCELLED, SUPERSEDED]
        proposedByRole:
          type: string
          enum: [customer, trader]
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        cancelledByRole:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, conversationId, startAt, status, proposedByRole, createdAt]
    CreateBookingRequest:
      type: object
      properties:
        startAt:
          type: string
          format: date-time
          description: Appointment start. Must be in the future.
        durationMinutes:
          type: integer
          enum: [30, 60, 90, 120, 180, 240, 480]
          description: Appointment length. Defaults to 60 for older clients.
        note:
          type: string
          maxLength: 300
          nullable: true
          description: Optional short note (e.g. "Initial visit and inspection").
      required: [startAt]
    BookingSlotsResponse:
      type: object
      properties:
        date:
          type: string
          description: The UK-local date requested (YYYY-MM-DD).
        durationMinutes:
          type: integer
        slots:
          type: array
          items:
            type: string
            format: date-time
          description: Available start instants (UTC) fitting the duration.
        hasWorkingHours:
          type: boolean
          description: Whether the trader has configured structured hours.
      required: [date, durationMinutes, slots, hasWorkingHours]
...
    BookingResponse:
      type: object
      properties:
        booking:
          $ref: "#/components/schemas/Booking"
      required: [booking]
...